+++ src/app/features/home/activities/activities.module.ts
@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from '../../../shared/shared.module';
-import { TransactionPageRoutingModule } from './activities-routing.module';
-import { TransactionPage } from './activities.page';
+import { ActivitiesPageRoutingModule } from './activities-routing.module';
+import { ActivitiesPage } from './activities.page';
 import { CaptureTransactionsComponent } from './capture-transactions/capture-transactions.component';
 import { NetworkActionOrdersComponent } from './network-action-orders/network-action-orders.component';
 
 @NgModule({
-  imports: [SharedModule, TransactionPageRoutingModule],
+  imports: [SharedModule, ActivitiesPageRoutingModule],
   declarations: [
-    TransactionPage,
+    ActivitiesPage,
     CaptureTransactionsComponent,
     NetworkActionOrdersComponent,
   ],
 })
-export class TransactionPageModule {}
+export class ActivitiesPageModule {}