+++ src/app/features/home/activities/activities.page.spec.ts
@@ -1,19 +1,19 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { SharedTestingModule } from '../../../shared/shared-testing.module';
-import { TransactionPage } from './activities.page';
+import { ActivitiesPage } from './activities.page';
 
-describe('TransactionPage', () => {
-  let component: TransactionPage;
-  let fixture: ComponentFixture<TransactionPage>;
+describe('ActivitiesPage', () => {
+  let component: ActivitiesPage;
+  let fixture: ComponentFixture<ActivitiesPage>;
 
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
-        declarations: [TransactionPage],
+        declarations: [ActivitiesPage],
         imports: [SharedTestingModule],
       }).compileComponents();
 
-      fixture = TestBed.createComponent(TransactionPage);
+      fixture = TestBed.createComponent(ActivitiesPage);
       component = fixture.componentInstance;
       fixture.detectChanges();
     })