+++ src/app/features/settings/go-pro/go-pro-media-item-detail-on-camera/go-pro-media-item-detail-on-camera.component.ts
@@ -51,10 +51,7 @@ export class GoProMediaItemDetailOnCameraComponent implements OnInit {
     if (allowed) {
       await this.startUploadToCapture();
 
-      if (
-        (await this.goProWiFiService.showTutorialForMobileDataOnlyApps()) ===
-        false
-      ) {
+      if (await this.goProWiFiService.showTutorialForMobileDataOnlyApps()) {
         this.showTutorialForMobileDataOnlyApps = true;
       }
     }