+++ src/app/features/settings/go-pro/services/go-pro-media.service.ts
@@ -2,12 +2,12 @@ import { HttpClient } from '@angular/common/http';
 import { Inject, Injectable } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import '@capacitor-community/http';
+import { Http } from '@capacitor-community/http';
+import { Capacitor } from '@capacitor/core';
 import {
-  Capacitor,
-  FilesystemDirectory,
+  Directory as FilesystemDirectory,
   FilesystemPlugin,
-  Plugins,
-} from '@capacitor/core';
+} from '@capacitor/filesystem';
 import { isPlatform } from '@ionic/core';
 import {
   detectFileTypeFromUrl,
@@ -18,7 +18,6 @@ import { FILESYSTEM_PLUGIN } from '../../../../shared/capacitor-plugins/capacito
 import { CaptureService } from '../../../../shared/capture/capture.service';
 import { blobToBase64 } from '../../../../utils/encoding/encoding';
 import { GoProFile } from '../go-pro-media-file';
-const { Http } = Plugins;
 
 @Injectable({
   providedIn: 'root',