+++ src/app/shared/version/version.service.ts
@@ -1,19 +1,22 @@
 import { Injectable } from '@angular/core';
-import { Plugins } from '@capacitor/core';
+import { App, AppInfo } from '@capacitor/app';
+import { isPlatform } from '@ionic/core';
 import { defer } from 'rxjs';
 import { map } from 'rxjs/operators';
 import packageInfo from '../../../../package.json';
-
-const { Device } = Plugins;
 
 @Injectable({
   providedIn: 'root',
 })
 export class VersionService {
-  readonly version$ = defer(() => Device.getInfo()).pipe(
+  readonly version$ = defer(async () => {
+    let appInfo: AppInfo = { version: '', build: '', id: '', name: '' };
+    if (isPlatform('hybrid')) appInfo = await App.getInfo();
+    return appInfo;
+  }).pipe(
     map(info => {
-      if (info.appVersion === '') return packageInfo.version;
-      return info.appVersion;
+      if (info.version === '') return packageInfo.version;
+      return info.version;
     })
   );
 }